Simplify type / app mapping
Currently you can assign multiple apps to a file type category *and* multiple
file extensions to a file type category. That means that given a file
extension and an app it's not always possible to determine the file category.
Further, these 2 degrees of freedom aren't necessary to correctly assign apps
to file extensions - allowing multiple apps per extension allows that.

So
* ensure that file extensions only appear once in the file category map
  * it turns out that deleting the custom categories fixed this issue
* Define the apps in a separate mapping from the category -> app mapping,
  since currently the same app is defined multiple times, once per each
  category.

This changes nothing in the resulting app mapping JSON other than the
ordering of some keys.