Update format for import specification templates
Makes this easier to deal with for xSV files. Since pandas drops data if
a row has more columns than the first row, manually parse the first line and
then pass the remainder of the file handle to pandas.

Tested with excel files - in that case pandas just duplicates the first cell
X times in the header structure where X is the maximum column count in the
data.