Set of changes to support multiple fixes to upload, as well as necessary code fixes for clarity. [PTV-1882]
All changed files are subject to black formatting, for uniformity and readability, and to remove linting errors.
The Path class in utils was renamed to StagingPath as it conflicts with Python's path, and forced a renaming of that to PyPath or PathPy.
Many changes in app.py, the home of all endpoints, and where half of the fixes occur.
Refactored much of the config-related code into a new file config.py
Hoisted many hard-coded values into constant-like variables at the module level.
Moved app startup logic into main (from __main__) as it is now the entrypoint for gunicorn
Applied stream-reading, md5-generation, text/binary recognition fixes to metadata.py.
Look for PR comment to explain the changes in more detail.