refactor entrypoint to use gunicorn and remove local dev support [PTV-1882]
The production entrypoint should not serve a dual purpose - local dev and prod runtime. The local dev support needs work anyway, as it relies on Visual Studio Code, which is out of date.
Gunicorn is used now as it helps ameliorate stuck aiohttp servers and relaunching crashed aiothttp servers. It's configuration uses env variables, all with defaults (which may need to thought about more deeply and/or tested.)