Development tools [PTV-1882]
Isolated development tools, to aid in development and testing within a container.
Although the host-based server and testing works, it requires installation of the precise same compiler version, which is not really possible cross-platform (though pretty close).
Also dependencies in linux (and in the specific version in deployment), do affect Python dependencies, so this reduces that variable greatly. (Of course, the way the service is run in a container in production will be different.)